finalized move
